requirements.txt
streamlit
langchain
langchain_openai
pinecone-client
datasets
pandas
athina
faiss-gpu